# api/openapi.yaml
openapi: 3.1.0
info:
  title: FlowPilot API
  version: 0.1.0
  description: >
    Contract-first REST API for FlowPilot. Paths and schemas are added slice
    by slice. PR1 adds Auth, PR2 adds the read-only user directory, PR3 adds
    Projects + BoardColumns, PR4 adds project Members, PR5 adds WorkItems.
    Hand-authored here is the source of truth; springdoc-openapi runs in CI
    only as a drift check against the running application (see
    OpenApiSpecExportTest + the oasdiff step in ci.yml). Paths below are
    absolute (include the /api prefix baked into every @RequestMapping) —
    there is no servers/context-path indirection, since the live
    springdoc-generated document has none either.
paths:
...
  /api/projects:
    get:
      summary: List projects visible to the caller
      operationId: listProjects
      description: >-
        Global administrators see every project; everyone else sees only
        projects they own. The list endpoint's ownership filter is unchanged
        by slice 4; single-resource reads (GET /projects/{id} and
        /board-columns) now also require admin/owner/member (see below).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      summary: Create a project
      operationId: createProject
      description: >-
        The caller becomes the project's owner. Seeds exactly 4 default
        BoardColumns ("Por hacer", "En progreso", "En revisión", "Terminado")
        atomically in the same transaction.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Project created, 4 BoardColumns seeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation failed, or start date is after estimated end date
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Missing or invalid access token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Code already belongs to another project
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/projects/{id}:
    get:
      summary: Get a project by id
      operationId: getProject
      description: >-
        Requires the caller to be a global administrator, the project's
        owner, or a ProjectMember of the project (closes the read-gap
        flagged in the slice-3 verify report: previously authentication-only).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No project with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      summary: Update a project's name/description
      operationId: updateProject
      description: >-
        Requires the PROJECT_EDIT_SETTINGS permission (role-permission
        matrix): a global administrator or the project's owner always has
        it, and by default so does the PROJECT_MANAGER project role
        (design's ProjectAuthorizationService).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation failed, or start date is after estimated end date
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Caller lacks the required permission for this project (admin, owner, or a matrix-granted role)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No project with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Code already belongs to another project, or the project was modified concurrently
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      summary: Delete a project
      operationId: deleteProject
      description: >-
        Requires the PROJECT_DELETE permission (role-permission matrix): a
        global administrator or the project's owner always has it;
        PROJECT_DELETE is granted to no project role by default, but an
        administrator may grant it to any role via the matrix.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Project deleted
        '403':
          description: Caller lacks the required permission for this project (admin, owner, or a matrix-granted role)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No project with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/projects/{id}/status:
    patch:
      summary: Transition a project's lifecycle status
      operationId: updateProjectStatus
      description: >-
        Requires the PROJECT_EDIT_SETTINGS permission (role-permission
        matrix), the same gate as updating the project itself.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectStatusUpdateRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '403':
          description: Caller lacks the required permission for this project (admin, owner, or a matrix-granted role)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No project with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/projects/{id}/board-columns:
    get:
      summary: List a project's Kanban columns, ordered by position
      operationId: listBoardColumns
      description: >-
        Read-only. Exactly 4 columns are seeded on project creation
        (design decision 9). Consumed by the future Kanban board slice.
        Requires the same admin/owner/member view gate as GET /projects/{id}.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Board columns ordered by position
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardColumnResponse'
        '403':
          description: Caller is not the owner, an administrator, or a member
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: No project with that id
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  parameters:
...
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    ProjectStatus:
      type: string
      enum: [PLANIFICACION, ACTIVO, PAUSADO, FINALIZADO, CANCELADO]
    ProjectCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        code:
          type: string
          maxLength: 50
          pattern: '^[A-Za-z0-9_-]*$'
        startDate:
          type: string
          format: date
        estimatedEndDate:
          type: string
          format: date
        technologies:
          type: string
          maxLength: 1000
        repositoryUrl:
          type: string
          maxLength: 500
          pattern: '^$|^https?://\S+$'
    ProjectUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        code:
          type: string
          maxLength: 50
          pattern: '^[A-Za-z0-9_-]*$'
        startDate:
          type: string
          format: date
        estimatedEndDate:
          type: string
          format: date
        technologies:
          type: string
          maxLength: 1000
        repositoryUrl:
          type: string
          maxLength: 500
          pattern: '^$|^https?://\S+$'
    ProjectStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/ProjectStatus'
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        ownerId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        code:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        estimatedEndDate:
          type: string
          format: date
          nullable: true
        technologies:
          type: string
          nullable: true
        repositoryUrl:
          type: string
          nullable: true
    BoardColumnResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        position:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT